Strangler extraction #11: Favorites tab + pane routing -> zxnu_favorites_pane.py
Move the Favorites feature out of MainWindow.__init__ as THREE builders
(the original block is interleaved with the pane-builder calls and tab
assembly that stay in the monolith), each called at its chunk''s
historical position:

* build_favorites_helpers — cross-pane record/toggle/save/navigate
  helpers (runs before the pane builders that consume host._fav_is/
  _fav_toggle; the three run_search closures they bind later are
  injected as forwarding lambdas).
* build_favorites_pane — the tab widget layer (getters, GalleryView,
  table, view selector, stack, addTab); also exposes the _fav_* getters
  the Unite! builders consume, re-bound to bare locals at the call site.
* build_favorites_ops — _fav_repopulate + the view-mode apply helper and
  the _pane_* per-pane Classic/Retro item-viewer routing layer; the
  closures'' local PygameSurfaceWidget variable ``host`` was renamed
  ``pg_host`` so it cannot shadow the builder''s host param.

zxnu_main.py 11,571 -> 10,925 lines. New module tripwired in
tests/test_pane_imports.py; full suite green.
